engine_rocks: move tests to their own module
